Support for setting both LISP heap and stack size
Add support for setting both the LISP heap and stack size (in Mb)
via the LISP_HEAP and LISP_STACK variables.  This will likely require
some downstream updates as previously the LISP_STACK variable was
used in setting the SBCL heap size.

This commit is necessary as a larger-than-default stack size
is required for performing differencing on some JavaScript projects
without blowing out the stack.  Unfortunately, I didn't see an
easy way to change the existing algorithm to use iteration or
be tail recursive although I could further explore these options if
this is desired instead.