Greatly enhance and test readme.py
Now the script uses `pytest` to make it easier to distinguish success
from failure, `difflib` to make it easier to read failure output, and
`tempfile` to make it easier to reproduce a failure.

The `test` directory now holds several Markdown examples, with
"normalized" outputs in `.txt` files and (for the examples that are
meant to be failures) the Lisp scripts that `readme.py` gives to
reproduce failure. These examples are all automatically checked in CI as
part of the new Python test suite.